pyproject.toml
[project]
name = "django-boilerplate"
version = "0.1.0"
description = "Django project template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-environ>=0.13.0",
    "django-health-check>=4.4.0",
    "django-mptt>=0.18.0",
    "django-tinymce>=5.0.0",
    "django-view-breadcrumbs>=2.5.1",
    "whitenoise>=6.12.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.3.0",
    "django-extensions>=4.1",
    "ipython>=9.13.0",
]
